+++ OVirtualMachine/LogicUnit.cpp
@@ -23,42 +23,42 @@ namespace O {
 	throw(std::exception());
     }
 
-    void LogicUnit::ga(long desc){
+    void LogicUnit::ga(long long desc){
     	throw(std::exception());
     }
 
-    void LogicUnit::mov(O::Memory::Registers destReg, long value) {
-        long* destination = _mem->GetRegisterAccess(destReg);
+    void LogicUnit::mov(O::Memory::Registers destReg, long long value) {
+        long long* destination = _mem->GetRegisterAccess(destReg);
         *destination = value;
     }
 
     void LogicUnit::mov(O::Memory::Registers destReg, O::Memory::Registers sourceReg) {
-        long value = *(_mem->GetRegisterAccess(sourceReg));
+        long long value = *(_mem->GetRegisterAccess(sourceReg));
         mov(destReg, value);
     }
 
-    void LogicUnit::mov(O::Memory::MemoryAddressDescriptor dest, long value) {
-        long* d = _mem->GetAccessByMemoryDescriptor(dest);
+    void LogicUnit::mov(O::Memory::MemoryAddressDescriptor dest, long long value) {
+        long long* d = _mem->GetAccessByMemoryDescriptor(dest);
         *d = value;
     }
 
     void LogicUnit::mov(O::Memory::MemoryAddressDescriptor dest, O::Memory::MemoryAddressDescriptor source) {
-        long* d = _mem->GetAccessByMemoryDescriptor(dest);
-        long* s = _mem->GetAccessByMemoryDescriptor(source);
+        long long* d = _mem->GetAccessByMemoryDescriptor(dest);
+        long long* s = _mem->GetAccessByMemoryDescriptor(source);
         *d = *s;
     }
 
     void LogicUnit::mov(O::Memory::Registers dest, O::Memory::MemoryAddressDescriptor source) {
-        long* d = _mem->GetRegisterAccess(dest);
+        long long* d = _mem->GetRegisterAccess(dest);
         *d = *_mem->GetAccessByMemoryDescriptor(source);
     }
 
     void LogicUnit::mov(O::Memory::MemoryAddressDescriptor dest, O::Memory::Registers source) {
-        long* d = _mem->GetAccessByMemoryDescriptor(dest);
+        long long* d = _mem->GetAccessByMemoryDescriptor(dest);
         *d = *_mem->GetRegisterAccess(source);
     }
 
-    void LogicUnit::add(O::Memory::Registers dest, long value) {
+    void LogicUnit::add(O::Memory::Registers dest, long long value) {
         GETREG(dest) += value;
     }
 
@@ -70,7 +70,7 @@ namespace O {
         GETREG(dest) += GETMAD(source);
     }
 
-    void LogicUnit::sub(O::Memory::Registers dest, long value) {
+    void LogicUnit::sub(O::Memory::Registers dest, long long value) {
         GETREG(dest) -= value;
     }
 
@@ -82,7 +82,7 @@ namespace O {
         GETREG(dest) -= GETMAD(source);
     }
 
-    void LogicUnit::add(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::add(O::Memory::MemoryAddressDescriptor dest, long long value) {
         GETMAD(dest) += value;
     }
 
@@ -94,7 +94,7 @@ namespace O {
         GETMAD(dest) += GETMAD(source);
     }
 
-    void LogicUnit::sub(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::sub(O::Memory::MemoryAddressDescriptor dest, long long value) {
         GETMAD(dest) -= value;
     }
 
@@ -110,7 +110,7 @@ namespace O {
         _mem->pop(O::Memory::Registers::eip);
     }
 
-    void LogicUnit::call(long value) {
+    void LogicUnit::call(long long value) {
         throw std::exception();
     }
 
@@ -123,7 +123,7 @@ namespace O {
         GETREG(O::Memory::Registers::eip) = _mem->GetIdByMAD(mad);
     }
 
-    void LogicUnit::cmp(O::Memory::Registers destReg, long value) {
+    void LogicUnit::cmp(O::Memory::Registers destReg, long long value) {
         auto p1 = GETREG(destReg);
         auto p2 = value;
         auto res = 0;
@@ -152,7 +152,7 @@ namespace O {
         cmp(destReg, p2);
     }
 
-    void LogicUnit::cmp(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::cmp(O::Memory::MemoryAddressDescriptor dest, long long value) {
         throw std::exception();
     }
 
@@ -168,7 +168,7 @@ namespace O {
         throw std::exception();
     }
 
-    void LogicUnit::move(O::Memory::Registers destReg, long value) {
+    void LogicUnit::move(O::Memory::Registers destReg, long long value) {
         auto flags = GETREG(Memory::Registers::flag);
         if((flags & 0b1) == 0b1){
             mov(destReg, value);
@@ -182,7 +182,7 @@ namespace O {
         }
     }
 
-    void LogicUnit::move(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::move(O::Memory::MemoryAddressDescriptor dest, long long value) {
         auto flags = GETREG(Memory::Registers::flag);
         if((flags & 0b1) == 0b1){
             mov(dest, value);
@@ -210,7 +210,7 @@ namespace O {
         }
     }
 
-    void LogicUnit::movg(O::Memory::Registers destReg, long value) {
+    void LogicUnit::movg(O::Memory::Registers destReg, long long value) {
         auto flags = GETREG(Memory::Registers::flag);
         if((flags & 0b10) == 0b10){
             mov(destReg, value);
@@ -224,7 +224,7 @@ namespace O {
         }
     }
 
-    void LogicUnit::movg(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::movg(O::Memory::MemoryAddressDescriptor dest, long long value) {
         auto flags = GETREG(Memory::Registers::flag);
         if((flags & 0b10) == 0b10){
             mov(dest, value);
@@ -252,7 +252,7 @@ namespace O {
         }
     }
 
-    void LogicUnit::movl(O::Memory::Registers destReg, long value) {
+    void LogicUnit::movl(O::Memory::Registers destReg, long long value) {
         auto flags = GETREG(Memory::Registers::flag);
         if((flags & 0b100) == 0b100){
             mov(destReg, value);
@@ -266,7 +266,7 @@ namespace O {
         }
     }
 
-    void LogicUnit::movl(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::movl(O::Memory::MemoryAddressDescriptor dest, long long value) {
         auto flags = GETREG(Memory::Registers::flag);
         if((flags & 0b100) == 0b100){
             mov(dest, value);
@@ -294,7 +294,7 @@ namespace O {
         }
     }
 
-    void LogicUnit::jmp(long value) {
+    void LogicUnit::jmp(long long value) {
         throw std::exception();
     }
 
@@ -306,7 +306,7 @@ namespace O {
         GETREG(O::Memory::Registers::eip) = _mem->GetIdByMAD(mad);
     }
 
-    void LogicUnit::jme(long value) {
+    void LogicUnit::jme(long long value) {
         throw std::exception();
     }
 
@@ -321,7 +321,7 @@ namespace O {
         }
     }
 
-    void LogicUnit::interrupt(long value) {
+    void LogicUnit::interrupt(long long value) {
         Interrupt inter = registeredInterrupts[value];
         inter.hInt(_mem->getMemPointer());
     }
@@ -349,14 +349,14 @@ namespace O {
         std::vector<Interrupt> newInterrupts = main(registeredInterrupts.size());
 
         for(auto i : newInterrupts){
-            std::vector<long> b = {17, 0, i.id, 0, 0,5,0,0,0,0};
+            std::vector<long long> b = {17, 0, i.id, 0, 0,5,0,0,0,0};
             _mem->LoadProgram(i.name, b);
         }
 
         registeredInterrupts.insert(registeredInterrupts.end(), newInterrupts.begin(), newInterrupts.end());
     }
 
-    void LogicUnit::mul(O::Memory::Registers destReg, long value) {
+    void LogicUnit::mul(O::Memory::Registers destReg, long long value) {
 
         GETREG(destReg) = GETREG(destReg) * value;
     }
@@ -365,7 +365,7 @@ namespace O {
         GETREG(destReg) = GETREG(destReg) * GETREG(sourceReg);
     }
 
-    void LogicUnit::mul(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::mul(O::Memory::MemoryAddressDescriptor dest, long long value) {
         GETMAD(dest) *= value;
     }
 
@@ -381,7 +381,7 @@ namespace O {
         GETMAD(dest) *= GETREG(source);
     }
 
-    void LogicUnit::div(O::Memory::Registers destReg, long value) {
+    void LogicUnit::div(O::Memory::Registers destReg, long long value) {
 
         GETREG(destReg) = GETREG(destReg) / value;
     }
@@ -390,7 +390,7 @@ namespace O {
         GETREG(destReg) = GETREG(destReg) / GETREG(sourceReg);
     }
 
-    void LogicUnit::div(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::div(O::Memory::MemoryAddressDescriptor dest, long long value) {
         GETMAD(dest) /= value;
     }
 
@@ -406,7 +406,7 @@ namespace O {
         GETMAD(dest) /= GETREG(source);
     }
 
-    void LogicUnit::mod(O::Memory::Registers destReg, long value) {
+    void LogicUnit::mod(O::Memory::Registers destReg, long long value) {
 
         GETREG(destReg) = GETREG(destReg) % value;
     }
@@ -415,7 +415,7 @@ namespace O {
         GETREG(destReg) = GETREG(destReg) % GETREG(sourceReg);
     }
 
-    void LogicUnit::mod(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::mod(O::Memory::MemoryAddressDescriptor dest, long long value) {
         GETMAD(dest) %= value;
     }
 
@@ -431,7 +431,7 @@ namespace O {
         GETMAD(dest) %= GETREG(source);
     }
 
-    void LogicUnit::AND(O::Memory::Registers destReg,long value) {
+    void LogicUnit::AND(O::Memory::Registers destReg,long long value) {
 
         GETREG(destReg) = GETREG(destReg) & value;
     }
@@ -440,7 +440,7 @@ namespace O {
         GETREG(destReg) = GETREG(destReg) & GETREG(sourceReg);
     }
 
-    void LogicUnit::AND(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::AND(O::Memory::MemoryAddressDescriptor dest, long long value) {
         GETMAD(dest) &= value;
     }
 
@@ -456,7 +456,7 @@ namespace O {
         GETMAD(dest) &= GETREG(source);
     }
 
-    void LogicUnit::OR(O::Memory::Registers destReg, long value) {
+    void LogicUnit::OR(O::Memory::Registers destReg, long long value) {
 
         GETREG(destReg) = GETREG(destReg) | value;
     }
@@ -465,7 +465,7 @@ namespace O {
         GETREG(destReg) = GETREG(destReg) | GETREG(sourceReg);
     }
 
-    void LogicUnit::OR(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::OR(O::Memory::MemoryAddressDescriptor dest, long long value) {
         GETMAD(dest) |= value;
     }
 
@@ -481,7 +481,7 @@ namespace O {
         GETMAD(dest) |= GETREG(source);
     }
 
-    void LogicUnit::addf(O::Memory::Registers destReg, long value) {
+    void LogicUnit::addf(O::Memory::Registers destReg, long long value) {
 
         float op1 = *((float*)_mem->GetRegisterAccess(destReg));
         float op2 = *((float*)&value);
@@ -500,7 +500,7 @@ namespace O {
         GETREG(destReg) = *((int*)&op1);
     }
 
-    void LogicUnit::addf(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::addf(O::Memory::MemoryAddressDescriptor dest, long long value) {
         float op1 = *((float*)_mem->GetAccessByMemoryDescriptor(dest));
         float op2 = *((float*)&value);
 
@@ -536,7 +536,7 @@ namespace O {
         GETMAD(dest) = *((int*)&op1);
     }
 
-    void LogicUnit::subf(O::Memory::Registers destReg, long value) {
+    void LogicUnit::subf(O::Memory::Registers destReg, long long value) {
 
         float op1 = *((float*)_mem->GetRegisterAccess(destReg));
         float op2 = *((float*)&value);
@@ -555,7 +555,7 @@ namespace O {
         GETREG(destReg) = *((int*)&op1);
     }
 
-    void LogicUnit::subf(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::subf(O::Memory::MemoryAddressDescriptor dest, long long value) {
         float op1 = *((float*)_mem->GetAccessByMemoryDescriptor(dest));
         float op2 = *((float*)&value);
 
@@ -591,7 +591,7 @@ namespace O {
         GETMAD(dest) = *((int*)&op1);
     }
 
-    void LogicUnit::mulf(O::Memory::Registers destReg, long value) {
+    void LogicUnit::mulf(O::Memory::Registers destReg, long long value) {
 
         float op1 = *((float*)_mem->GetRegisterAccess(destReg));
         float op2 = *((float*)&value);
@@ -610,7 +610,7 @@ namespace O {
         GETREG(destReg) = *((int*)&op1);
     }
 
-    void LogicUnit::mulf(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::mulf(O::Memory::MemoryAddressDescriptor dest, long long value) {
         float op1 = *((float*)_mem->GetAccessByMemoryDescriptor(dest));
         float op2 = *((float*)&value);
 
@@ -646,7 +646,7 @@ namespace O {
         GETMAD(dest) = *((int*)&op1);
     }
 
-    void LogicUnit::divf(O::Memory::Registers destReg, long value) {
+    void LogicUnit::divf(O::Memory::Registers destReg, long long value) {
 
         float op1 = *((float*)_mem->GetRegisterAccess(destReg));
         float op2 = *((float*)&value);
@@ -665,7 +665,7 @@ namespace O {
         GETREG(destReg) = *((int*)&op1);
     }
 
-    void LogicUnit::divf(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::divf(O::Memory::MemoryAddressDescriptor dest, long long value) {
         float op1 = *((float*)_mem->GetAccessByMemoryDescriptor(dest));
         float op2 = *((float*)&value);
 
@@ -701,7 +701,7 @@ namespace O {
         GETMAD(dest) = *((int*)&op1);
     }
 
-    void LogicUnit::cmpf(O::Memory::Registers destReg, long value) {
+    void LogicUnit::cmpf(O::Memory::Registers destReg, long long value) {
         float p1 = *((float*)_mem->GetRegisterAccess(destReg));
         float p2 = *((float*)&value);
         auto res = 0;
@@ -730,7 +730,7 @@ namespace O {
         cmp(destReg, p2);
     }
 
-    void LogicUnit::cmpf(O::Memory::MemoryAddressDescriptor dest, long value) {
+    void LogicUnit::cmpf(O::Memory::MemoryAddressDescriptor dest, long long value) {
         throw std::exception();
     }
 
@@ -747,7 +747,7 @@ namespace O {
     }
 
     void LogicUnit::AddNewInterrupt(std::string name, InterruptHandler interrupt) {
-        std::vector<long> b = {17, 0, (long)registeredInterrupts.size(), 0, 0,5,0,0,0,0};
+        std::vector<long long> b = {17, 0, (long long)registeredInterrupts.size(), 0, 0,5,0,0,0,0};
         _mem->LoadProgram(name, b);
 
         Interrupt inter;
